Eliminate or suppress warnings.

+++ plugins/i3s/lepcc/src/Intensity.cpp
@@ -165,7 +165,7 @@ ErrCode Intensity::GetBlobSize(const Byte* pByte, int64 bufferSize, uint32& blob
 
   TopHeader refHd;
   Header1 hd1;
-  if (bufferSize < sizeof(refHd) + sizeof(hd1.blobSize))
+  if (bufferSize < (int64)(sizeof(refHd) + sizeof(hd1.blobSize)))
     return ErrCode::BufferTooSmall;
 
   if (0 != memcmp(pByte, refHd.fileKey, refHd.FileKeyLength()))    // file key
